+++ demo/fake_diffusion/run.py
@@ -3,9 +3,10 @@ import numpy as np
 import time
 
 def fake_diffusion(steps):
+    rng = np.random.default_rng()
     for i in range(steps):
         time.sleep(1)
-        image = np.random.random((600, 600, 3))
+        image = rng.random(size=(600, 600, 3))
         yield image
     image = np.ones((1000,1000,3), np.uint8)
     image[:] = [255, 124, 0]