+++ gradio/blocks.py
@@ -1325,7 +1325,7 @@ Received inputs:
                     if input_id in state:
                         block = state[input_id]
                     inputs_cached = processing_utils.move_files_to_cache(
-                        inputs[i], block
+                        inputs[i], block, add_urls=True
                     )
                     if getattr(block, "data_model", None) and inputs_cached is not None:
                         if issubclass(block.data_model, GradioModel):  # type: ignore
@@ -1449,11 +1449,15 @@ Received outputs:
                         raise InvalidComponentError(
                             f"{block.__class__} Component with id {output_id} not a valid output component."
                         )
+                    if output_id in state:
+                        block = state[output_id]
                     prediction_value = block.postprocess(prediction_value)
+
                 outputs_cached = processing_utils.move_files_to_cache(
                     prediction_value,
                     block,  # type: ignore
                     postprocess=True,
+                    add_urls=True,
                 )
                 output.append(outputs_cached)