+++ gradio/routes.py
@@ -17,7 +17,6 @@ import mimetypes
 import os
 import posixpath
 import secrets
-import shutil
 import tempfile
 import threading
 import time
@@ -54,6 +53,7 @@ from gradio.context import Context
 from gradio.data_classes import ComponentServerBody, PredictBody, ResetBody
 from gradio.exceptions import Error
 from gradio.oauth import attach_oauth
+from gradio.processing_utils import add_root_url
 from gradio.queueing import Estimation
 from gradio.route_utils import (  # noqa: F401
     FileUploadProgress,
@@ -63,6 +63,7 @@ from gradio.route_utils import (  # noqa: F401
     GradioUploadFile,
     MultiPartException,
     Request,
+    move_uploaded_files_to_cache,
 )
 from gradio.state_holder import StateHolder
 from gradio.utils import (
@@ -114,12 +115,6 @@ templates = Jinja2Templates(directory=STATIC_TEMPLATE_LIB)
 templates.env.filters["toorjson"] = toorjson
 
 client = httpx.AsyncClient()
-
-
-def move_uploaded_files_to_cache(files: list[str], destinations: list[str]) -> None:
-    for file, dest in zip(files, destinations):
-        shutil.move(file, dest)
-
 
 file_upload_statuses = FileUploadProgress()
 
@@ -315,20 +310,18 @@ class App(FastAPI):
         def main(request: fastapi.Request, user: str = Depends(get_current_user)):
             mimetypes.add_type("application/javascript", ".js")
             blocks = app.get_blocks()
-            root_path = (
-                request.scope.get("root_path")
-                or request.headers.get("X-Direct-Url")
-                or ""
+            root = route_utils.get_root_url(
+                request=request, route_path="/", root_path=app.root_path
             )
             if app.auth is None or user is not None:
                 config = app.get_blocks().config
-                config["root"] = route_utils.strip_url(root_path)
+                config = route_utils.update_root_in_config(config, root)
             else:
                 config = {
                     "auth_required": True,
                     "auth_message": blocks.auth_message,
                     "space_id": app.get_blocks().space_id,
-                    "root": route_utils.strip_url(root_path),
+                    "root": root,
                 }
 
             try:
@@ -359,14 +352,12 @@ class App(FastAPI):
         @app.get("/config/", dependencies=[Depends(login_check)])
         @app.get("/config", dependencies=[Depends(login_check)])
         def get_config(request: fastapi.Request):
-            root_path = (
-                request.scope.get("root_path")
-                or request.headers.get("X-Direct-Url")
-                or ""
-            )
             config = app.get_blocks().config
-            config["root"] = route_utils.strip_url(root_path)
-            return config
+            root = route_utils.get_root_url(
+                request=request, route_path="/config", root_path=app.root_path
+            )
+            config = route_utils.update_root_in_config(config, root)
+            return ORJSONResponse(content=config)
 
         @app.get("/static/{path:path}")
         def static_resource(path: str):
@@ -580,6 +571,10 @@ class App(FastAPI):
                     content={"error": str(error) if show_error else None},
                     status_code=500,
                 )
+            root_path = route_utils.get_root_url(
+                request=request, route_path=f"/api/{api_name}", root_path=app.root_path
+            )
+            output = add_root_url(output, root_path, None)
             return output
 
         @app.get("/queue/data", dependencies=[Depends(login_check)])
@@ -588,6 +583,9 @@ class App(FastAPI):
             session_hash: str,
         ):
             blocks = app.get_blocks()
+            root_path = route_utils.get_root_url(
+                request=request, route_path="/queue/data", root_path=app.root_path
+            )
 
             async def sse_stream(request: fastapi.Request):
                 try:
@@ -633,6 +631,7 @@ class App(FastAPI):
                                 "success": False,
                             }
                         if message:
+                            add_root_url(message, root_path, None)
                             yield f"data: {json.dumps(message)}\n\n"
                             if message["msg"] == ServerMessage.process_completed:
                                 blocks._queue.pending_event_ids_session[